# docs/api-documentation/openapi-spec-api-gateway.yaml
openapi: 3.0.3
info:
  title: Food Delivery System API Gateway
  description: |
    API Gateway untuk Food Delivery System Microservices dengan JWT Authentication
    
    ## Features
    - JWT Authentication dengan demo users
    - Service proxy ke 5 microservices
    - CORS support untuk frontend
    - Comprehensive error handling
    - Health check endpoints
    - Swagger UI documentation
    
    ## Demo Credentials
    - **Admin**: username: `admin`, password: `admin123`
    - **User**: username: `user`, password: `user123`
    
    ## Authentication
    Semua protected endpoints memerlukan Bearer token yang obtained via `/auth/login`
  version: 1.0.0
  contact:
    name: Kelompok 01
    email: food.delivery@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.fooddelivery.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login User
      description: Login user dan dapatkan JWT access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register User
      description: Register user baru
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Invalid request data or username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/user-service/{path}:
    parameters:
      - name: path
        in: path
        required: true
        schema:
          type: string
          example: api/users
    get:
      tags: [User Service]
      summary: Proxy GET requests to User Service
      description: Forward GET requests to User Service (Port 5001)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Success response from User Service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: User Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token obtained from /auth/login
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          example: admin123
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          $ref: '#/components/schemas/UserInfo'
        message:
          type: string
          example: Login successful
    RegisterRequest:
      type: object
      required: [username, password, email]
      properties:
        username:
          type: string
          example: newuser
        password:
          type: string
          example: password123
        email:
          type: string
          format: email
          example: newuser@example.com
    RegisterResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: User registered successfully
...
    UserInfo:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: admin
        email:
          type: string
          example: admin@fooddelivery.com
        role:
          type: string
          enum: [admin, user]
          example: admin
    ServiceResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Service not found
        message:
          type: string
          example: The requested service is not available
...